requirements.txt
pypdf>=4.1.0
langchain>=0.1.12
langchain-core>=0.1.32
langchain-community>=0.0.28
langchain-google-genai>=0.0.11
qdrant-client>=1.8.0
openpyxl>=3.1.2
docx2txt>=0.8
python-multipart>=0.0.9
unstructured>=0.11.8
langchain-groq>=0.0.1
psutil>=5.9.8
networkx>=3.2.1
PyPDF2
python-dotenv
flask
flask_cors